+++ R/launch_shiny.R
@@ -22,10 +22,23 @@
 #' @param fun The function to run as alternative to start the Shiny application
 #' locally. It is better to fully specify it (`package::function`), and it
 #' should take one argument which is the application name in `app =`.
-#' @param alt1 Alternate text to display at the bottom of the screenshot. If
+#' @param alt1 Alternate text to display at the bottom of the screenshot when
+#' nothing is provided for `fun =`. If
 #' `NULL`, nothing is displayed below the screenshot.
 #' @param alt2 Alternate text to display at the bottom of the screenshot in case
 #' `fun =` is provided.
+#' @param toc Entry to use in the exercises table of content (`NULL` if no
+#' entry, `""` for a default entry based on `toc.def =`).
+#' @param toc.def Text for a default toc entry using [glue()] syntax for
+#' replacement, e.g., `{app}`.
+#' @param run.url The URL to use to start the Shiny application in RStudio
+#' server in the SciViews Box. It should generally end with `?runrcode=`, and
+#' the R code to execute will be appended to it from `run.arg =`.
+#' @param run.cmd The command to use to launch the Shiny application in RStudio.
+#' @param run.arg The URL encoded version of `run.cmd =`.
+#' @param app.img The image to display in front of the toc entry
+#' @param app.link The link when the image is clicked (sends to an help page
+#' about Shiny applications).
 #' @param ... Not used here, but it allows to add more arguments used by the
 #' screenshot addin, like `delay =`, `offsetx =` or `offsety =`, see
 #' [webshot_shiny()].
@@ -42,15 +55,27 @@
 #' # TODO...
 launch_shiny <- function(url, app = basename(url),
 imgdir = "images/shinyapps", img = paste0(imgdir, "/", app, ".png"),
-createimg = TRUE, width = 780, height = 500, fun,
+createimg = TRUE, width = 780, height = 500, fun = NULL,
 alt1 = "*Click to start the Shiny application.*",
-alt2 = "*Click to start or run `{fun}(\"{app}\")`.*", ...) {
-  # Default value for app
-  if (missing(app))
-    app <- basename(url)
+alt2 = "*Click to start or [run `{run.cmd}`]({run.url}{run.arg}).*",
+toc = "", toc.def = "Shiny application {app}",
+run.url = "start_rstudio.html?runrcode=", run.cmd = glue("{fun}(\"{app}\")"),
+run.arg = URLencode(run.cmd, reserved = TRUE),
+app.img = "images/list-app.png", app.link = "shiny_app", ...) {
+  if (!is.null(toc)) {
+    # Add an entry in the ex_toc
+    ex_toc <- getOption("learndown_ex_toc", "")
+    if (toc == "") {
+      # Use default text
+      toc <- glue::glue(toc.def)
+    }
+    ex_toc <- paste0(ex_toc, "\n",
+      glue::glue("- [![app]({app.img})]({app.link}) [{toc}](#{app})"))
+    options(learndown_ex_toc = ex_toc)
+  }
 
   # Use alt2 if fun is provided, otherwise, use alt1
-  if (!missing(fun) && !is.null(fun) && fun != "") {
+  if (!is.null(fun) && fun != "") {
     alt <- glue(alt2)
   } else {
     alt <- glue(alt1)
@@ -81,8 +106,8 @@ alt2 = "*Click to start or run `{fun}(\"{app}\")`.*", ...) {
   }
 
   # Construct the HTML code to be included
-  res <- glue("\n[]{{#{app} }}<img onclick=\"launchApp('{app}', '{url}');\" src=\"{img}\" width=\"{width}\" height=\"{height}\" class=\"shiny-img\" id=\"img{app}\"/>
-<iframe width=\"{width}\" height=\"{height}\" frameborder=\"0\" scrolling=\"auto\" style=\"display:none\" class=\"shiny-app\" id=\"{app}\"></iframe>\n")
+  res <- glue("\n[]{{#{app} }}<img onclick=\"launchApp('{app}shiny', '{url}');\" src=\"{img}\" width=\"{width}\" height=\"{height}\" class=\"shiny-img\" id=\"img{app}shiny\"/>
+<iframe width=\"{width}\" height=\"{height}\" frameborder=\"0\" scrolling=\"auto\" style=\"display:none\" class=\"shiny-app\" id=\"{app}shiny\"></iframe>\n")
   if (!is.null(alt))
     res <- paste0(res, "\n", alt, "\n")
   res